client creates map from WorldMapSchema

+++ server/src/schemas/WorldMapSchema.ts
@@ -4,5 +4,5 @@ export class WorldMapSchema extends Schema {
   @type("string") mapName: string = "";
   @type("number") mapWidth: number = 0;
   @type("number") mapHeight: number = 0;
-  @type(["uint8"]) tileIndices: ArraySchema<number> = new ArraySchema<number>();
+  @type(["uint16"]) tileIndices: ArraySchema<number> = new ArraySchema<number>();
 }
